recursive (generic) QFT example (#42)
this PR implements a generic QFT using recursion and for loop. While we
are missing loop unroll, which should be relatively easy to support for
`scf.For` statement later. This example currently runs via
`QrackInterpreter`.

```python
@qasm2.extended
def qft(qreg: qasm2.QReg, n: int):
    if n == 0:
        return qreg

    qasm2.h(qreg[0])
    for i in range(1, n):
        qasm2.cu1(qreg[i], qreg[0], 2 * math.pi / 2**i)
    qft(qreg, n - 1)
    return qreg


@qasm2.extended
def main():
    return qft(qasm2.qreg(3), 3)
```

PS. well maybe this is a solution to #49 for QASM2 if we restrict
`qasm2.main` and `qasm2.gate`, just to make example look nicer. We will
need to worry about how to rewrite this to QASM2 tho or we just wait for
the IR serialization then we don't need QASM2.

This PR also includes several fixes to the address analysis and
`PyQrack` target.

CI failure due to missing Kirin 0.11 in the registry, I'm going to tag
kirin soon.