log-depth GHZ tutorial subsequent revisions/polish (#130)
I just wanted to fix some of the math after a discussion with @kaihsin
last week as well as add some nice admonitions using a "trick" from the
bloqade-python docs to get jupytext to render mkdocs elements. Also
tried to fix the wording in the note at the bottom of the tutorial
clarifying the validity of args vs. no args to the main qasm2 function.