Heuristic noise model (#169)
This PR adds a rewrite pass that inserts noise into a a qasm2 kernel. 

The idea here is to provide an interface that allows a user to describe
how a CZ entangling gate leads to extra noise do to the reconfiguration
of the atoms in order to do the entangling gate. This is on top of the
built-in rewrite that inserts the noise from gates.

NOTE: This model is not guaranteed to be supported long-term in bloqade.
We will be
moving towards a more general approach to noise modeling in the future.

---------

Co-authored-by: Xiu-zhe (Roger) Luo <rogerluo.rl18@gmail.com>