Fixing bug in `UOpToParallelRule` (#186)
The rewrite works by picking an arbitrary gate instruction in the gate
instructions that have been marked to be merged and then moving the
qubits to be close to that node. That is problematic in situations where
a register is defined after that node. I attempted to fix this by moving
all dependencies, but this didn't seem to work and is very complicated
to check in different situations, e.g., when the SSA values are block
arguments. So, instead, I decided that the solution is to just sort the
gate instructions in order of when they appear in the IR, making their
global ordering in the DAG. In that way, I can do the parallel rewrite
on the gate that appears last in the order of gates, ensuring that all
ssa dependencies are indeed available for the rewrite.

EDIT: The bug manifests as broken SSA references so I am running the
address analysis causes the tests to raise exceptions for catching this
particular error.