Fix bug in UOpToParallel (#190)
The previous patch was incorrect the gates can't move to the last gate
in the group becase that could potentially move a gate through another
gate it doesn't commute with, the opposite isn't true because the
operations occur in order of when they occur first in the IR. One issue
is that if you move the gates earlier, you need to move the references
to the qubits and also the registers as well. As such, I have added an
extra rewrite that moves all register definitions to the top of their
block. This should be fine in most cases since I optimize per block
anyway.