Allow gate noise to be inserted even if Address analysis fails for CZ gate. (#191)
In some circumstances, the address analysis gives up on providing a
global qubit address. This PR enables the heuristic noise to inject the
gate noise even if the address analysis doesn't give a global qubit
address. This is possible because the gate noise doesn't require global
qubit information it just needs a qubit reference to do the rewrite.