Adding interface for single zone heuristic noise models (#196)
This PR sketches an interface for building a heuristic noise model for a
single zone layout. The only requirement is implementing a method that
calculates the time for reconfiguration of the atoms to do the CZ gates
and forwards that to that duration to the idle error of the qubits that
are not moved.