Fix default arguments for PyQrack.  (#223)
I'm gonna open an issue on PyQrack repo about this it would be good to
know what are good flags to pass into the `QrackSimulator` object.
Currently, some configurations of options will cause a runtime error
when executing the simulation instead of a value error when constructing
the object. I'll add validation on the arguments passed into the
`QrackSimulator` object as we slowly figure out the bad behavior.