Change Clifford String to Pauli String (#224)
Per yesterday's meeting Scientific Software meeting agreement, the
Clifford string operator should be restricted to just Pauli's.

@weinbe58 @Roger-luo I know there were several methods mentioned to
translating the string into an actual `OpType` that can be used in other
statements, ranging from lowering, canonicalization rewrite, etc. What
is the _exact_ approach that should be taken for this? I'd like to open
an issue to keep track of the infrastructure required for this.

cc: @cduck @kaihsin