Adding noise statements to squin (#211)
Some initial friction points/anticipated changes I can immediately see
in translating this to stim

- The original `PauliChannel` definition from @kaihsin can take a
variable number of float arguments, but in the stim dialect there are
explicit fields for each probability of a 1Q/2Q pauli operator being
applied. My feeling is that this should be reflected on the squin side
as just something like `1QPauliChannel` and `2QPauliChannel` with
explicit fields. I don't think we'd lose any generality here along with
the added benefit that it would be much harder for a user to misinput
information (the default definition assumes you plug in either 3 or 15
values via varargs but it seems to easy to make a mistake here)
- ~~If we DO want varargs I'm not quite sure how to nicely feed this to
the kirin `@wraps` decorator~~
- From #200 , I see it could be possible to not have to provide an
operator for `basis` and instead use a `CliffordString` (although I'd
have to restrict CliffordString even further, to just the Pauli
Operators. Would it be worth having something like a `PauliString`? Or
is that silly?)

cc: @weinbe58 @Roger-luo @kaihsin

---------

Co-authored-by: Phillip Weinberg <weinbe58@gmail.com>