simplify loads (#240)
address #238 one question tho, was there a particular reason why the
`loads` and `loadfile` was living inside the lowering target? @weinbe58
I think if not would be nice to move them out. This is not an object we
expect users to use directly cuz it would be too confusing without
knowing about lowering process.