cleanup loads/loadfile in lowering (#242)
follow up #240 I removed the `returns` option cuz it seems no one is
using it. We can add something back if there is an actual user asking
for it.

I'm not sure why it was copy pasting `QASM2.run` method in the old
`loads` implementation, maybe there was some issue with the lowering so
this was a workaround? @weinbe58