Emit a cirq.Circuit from a squin kernel (#311)
There's still quite a few things to do here:

* Some of the operator statements (e.g. `Kron`) don't have a 1:1 mapping
to a cirq gate. They probably need a "runtime", since they are easy
enough to apply to qubits. Alternatively, we could use custom gates in
cirq, but I'm not much of a fan.
* Methods for the noise statements are still missing.
* There is some very hacky stuff I'm doing right now in order to obtain
the `BlockArguments` and `ReturnValues` in the frame for nested kernels.
There's probably a better way to do this, but I don't know how (yet).

---------

Co-authored-by: Phillip Weinberg <weinbe58@gmail.com>