adding Wired statement for regionalized wire semantics (#340)
The idea here is to replace phase out `wrap` and `unwrap` with
regionalized semantics. In this case the inputs to the `wired` statement
are the qubits to be converted into wires as block arguments that have
been unwrapped and when executing the region the last wires are
automatically wrapped back to their respective qubits.