Add heuristic noise model in cirq (#369)
Credits to @tcochran-quera and @lamq317 for the implementation.

For now, this is mostly just a simple port of the internally developed
heuristic noise model in cirq.
There are some potential improvements I see:

- The `get_two_zoned_noisy_circ` could use a custom class that inherits
from `cirq.NoiseModel` to organize the code and make it a bit cleaner.
- The `utils.py` could also use some clean up. There might still be some
functions that are actually not needed that I missed. Also we may want
to think about moving the functions to classes to organize the code a
bit more.

I didn't do any of that so far for a few reasons

- Mainly, this would break our internally developed notebooks, without
adding any benefits besides a bit of a cleaner codebase (no new features
or performance gains).
- We can still do it in the future and it's often useful to make changes
one at a time.


@tcochran-quera @lamq317 @jon-wurtz please have a look at this PR and
try it out. Let me know if you find things that can be improved /
cleaned up / unified with existing functionality / removed.


One side note: this PR also adds the lowering to `squin`, meaning you
can use `squin.cirq.load_circuit` in order to obtain a squin kernel from
a (noisy) circuit.


**Edit:** For posterity, adding the summary of the RFC posted in a
comment below here as well

# RFC summary

I'd like to summarize what we've discussed so far. Hopefully, this lets
us conclude the "RFC" part of this PR and move on to finalizing things.

- **General API**: There will be a single, high-level transformation
function `bloqade.cirq_utils.noise.transform_circuit(circuit, model)`,
which essentially does two things:
1. Converts the circuit to one using the native gate set only using
`cirq.optimize_for_target_gateset`.
2. Then applies the noise from the model (using what is essentially
`circuit.with_noise(model)`). Idle qubits that are part of the
non-native circuit will still be subjected to the move noise (this was
the discussion on init noise we had).
- **Two-zone noise**:
- The two-zone noise transformation is refactored into a
`cirq.NoiseModel`. This allows both us and the user to use
`circuit.with_noise` with the respective model. It should be noted,
however, that the noise models are only compatible with native circuits.
Also, there might be some edge cases due to initialization noise. This
needs to be clearly documented in the respective doc strings.
- The utility functions for the two-zone noise model are mostly left
as-is. However, I'd consider them as internal helper functions only, so
we should be able to make changes later on without breaking things for
users.
- The previously used function `get_two_zoned_noisy_circ` will be
removed. Users should use `transform_circuit` instead.
- **One-zone noise**:
- The interleaving is only part of one-zone noise models and as such
will be moved into those noise models. Again, this makes it possible for
the user to also just use `circuit.with_noise` so long as they keep the
restriction to native gates in mind.
- The `transform_to_noisy_one_zone_circuit` will also be removed and
replaced by `transform_circuit`.

---------

Co-authored-by: tcochran-quera <tcochran@quera.com>