Add tests from other infrastructure + enforce structure (#411)
Some of the tests here were contributed by @liupengy19 in another repo
that uses infrastructure based on SquinToStim. While they were
originally "higher up" in the pipeline the root cause of the problems
was in SquinToStim (as opposed to that other infrastructure).

Therefore I thought it better to move the tests here, closer to what
actually caused them to fail in the first place. I've also opted to put
the tests in their proper locations, as opposed to just have one file
for them from @kaihsin 's original structure (understandably faster but
makes it more confusing as to why the test is where it is).

---------

Co-authored-by: Kai-Hsin Wu <kaihsinwu@gmail.com>