Adapt SquinToStim to use new MeasureIDAnalysis (#414)
This is dependent on PR #413 (Please review that one first) but the new
SquinToStim pass should now be able to generate the correct record
indices. A couple of comments in terms of the changes:

- I don't see a need to generate a GetRecord unless it NEEDS to be
there. The only time this is a reality is when IfsToStim sees a
compatible boolean condition, in which case a GetRecord should be
generated.
- There was one pre-existing test with regards to SquinToStim that
failed with the new changes, but this was because the record indexes
were historically incorrect. The test in question is `simple_if_rewrite`
in `test/stim/passes/test_squin_meas_to_stim.py`

```
# The test's structure (in pseudocode) is:

q = qubits(4)
ms = measure(q)

if ms[0]
    # apply some gates that can be translated to feedforward

if ms[1]
    # same as above
   
measure(q) # another measurement
```

Because of #398 the global number of measures was used and the indexes
generated were `-7` and `-8`. I think this is still technically correct
(they are first - and oldest - measurements that occur) but the more
correct behavior is that at the time of the scf.IfElse you only have 4
measurements total so the correct index generated should be -4 and -3.

---------

Co-authored-by: Kai-Hsin Wu <kaihsinwu@gmail.com>