Implement stdlib shorthands for direct application of squin gates (#415)
This supersedes #393.

There's now `squin.gate` and `squin.parallel`, which allow the user to
apply simple gates quickly.
Not sure about the naming, but `gate` is basically `apply` and
`parallel` is `broadcast`.

For example

```python
@squin.kernel
def main():
    q = squin.qubit.new(4)
    squin.gate.h(q[0])
    squin.parallel.x(q)
```

One implementation detail here: Due to the `ApplyDesugar` pass, we need
to inline the `gate` calls since it uses `apply`. If we don't inline
those, the desugar pass doesn't do its magic and we are stuck with an
`apply_any`. This can go once we fully remove the syntax `apply(op: Op,
qubits: IList[Qubit])`. For now, however, I also inlined the `parallel`
stuff to ensure they behave the same.