Support Slice propogation in MeasurementIDAnalysis (#438)
I was working on a more advanced unit test for SquinToStim when I ran
into this problem.

If you use slice on a list of measurements the analysis doesn't go
through. The solution was to just carry over the logic I figured out for
the address analysis. However I cooked up a case where even this is
insufficient and figured out the simplest addition to the logic to amend
this case.


Basically both measure and address have something like the following:
```
# container could be AddressTuple or MeasureIDTuple
getitem(container, slice) -> tuple(....)
```
Where tuple is literally a plain Python tuple. This doesn't present any
problems until you slice again:

```
getitem(tuple, slice) -> ???
```
in which case the method table impl is blind to how to handle a raw
tuple. I believe the more correct way is that slice should generate
ANOTHER container type instead of a naked python tuple.

---------

Co-authored-by: Kai-Hsin Wu <kaihsinwu@gmail.com>