Make all GeminiOneZoneModels use correlated noise (#459)
Implements #448.

The test on the noisy GHZ fidelity now fails. On the one hand, of
course, since the fidelities did change. On the other hand, however, we
now somehow get a singular matrix for two qubits which produces a
corresponding `np.LinAlg` warning and results in a fidelity > 1. FWIW,
checking manually you can see that the matrices are pretty close, so the
issue really seems to be with the way the fidelity is calculated.

What's also kind of fishy: the fidelity for three qubits increased
(slightly). Before, we had: 0.9518, now it's 0.957. @tcochran-quera
@lamq317 did you see something similar when comparing the one-zone model
with and without correlated noise or is there something wrong here?


You can now also pass in the correlated rates either as dictionary or as
array

```python
model = GeminiOneZoneModel(cz_paired_correlated_rates = np.array(...))
model = GeminiOneZoneModel(cz_paired_error_probabilities = {"II": 0.9, "XX": ...})
```

Not sure about the naming though.

---------

Co-authored-by: Luis Martinez <lamq317@gmail.com>
Co-authored-by: Tyler A Cochran <TCochran@quera.com>