Implement methods for squin noise statements (#478)
Implements #476 .

In general, I think we need to keep the `StochasticUnitaryChannel`
statement, as there is no other way to faithfully lower cirq's
`AsymmetricDepolarizingChannel`. We may want to consider renaming the
statement, however, since it's really only needed for that now.

We can still discuss whether we want to keep the rewrite, I'd just have
to revert the last commit. But I wanted to point out that this way we
actually reduce the amount of code.

This might supersede #469.