Ensure no overlapping operations when emitting broadcast to cirq (#491)
There's a bug here when trying to broadcast composite operators such as
a multiplication. The problem is that in the cirq emit, a multiplication
is emitted as a sequence of gates on the same set of qubits, which can't
be added into the same moment (which is what broadcast emit did before).
Instead, we want to have a broadcast of multiplication emitted as a
sequence of moments.

Fortunately, we can just let cirq handle this by directly appending to
the circuit instead of creating moments.

---------

Co-authored-by: Xiu-zhe (Roger) Luo <rluo@quera.com>