Upgrading PyQrack dependency and workflows.  (#504)
Apparently, in the latest version they have added a flag to disable
OpenCL, which I think will make supporting pyqrack a lot easier. I want
to try this out and see if it works!

EDIT: doesn't work, however, I can specify a fine-grained platform which
should fix the issues with which pyqrack to install.