Implement clifford dialect (#503)
This implements the clifford dialect under squin, as discussed in #494.

Consistent with #502, all statements take a list of qubits as arguments.
We don't expose the wrappers, instead, there is a `simple` stdlib and a
`broadcast` stdlib, which contains kernel functions that provide the
user-facing API.

Some open things:
* I'm re-exporting the standard library functions on the top-level of
squin. This is for usability, so the user can just write e.g.
`squin.x(q[0])`, which calls the `stdlib.simple.x` method. Similarly,
you can call `squin.broadcast.x(q)`. The only downside of this, IMO, is
that things might be a bit confusing when we backport this. But we can
also remove the export and add it back later when we are doing the
breaking refactor.
* There are a few statements, which I'm not sure we need, so I skipped
them for now. Once we settle on which statements should be added, I'll
add the rest as stdlib functions. The statements are:
  * `Shift`
  * `U3`
* `Rot` - this one is actually part of the native dialect, but I'm not
sure we need it in clifford, since I already added `Rx` and `Ry`.
@weinbe58 what do you think?


**Edit**: The way I export stuff here is somewhat related to
https://github.com/QuEraComputing/bloqade/issues/287. As @Roger-luo
suggests there, we could also move the stdlib here to `gate`, but that
is breaking since e.g. `gate.x` is already taken as of now.