Inline function invokes when emitting a cirq circuit (#533)
Closes #527 .

@weinbe58 I originally tried to simply keep a function from being cached
if it allocates new qubits. In principle, I had all the information
there via the `sub_frame.qubit_index`, which, if incremented, would
reflect that there have been qubits allocated.

However, when I had that, there was still _something_ off with the
caching: it would sometimes produce a circuit that applied two Hadamard
gates to the first qubit and none to the third one. The odd part was
that this only occurred sometimes, other times the circuit looked fine.
I'm not entirely sure what the cause was here, but probably something
with caching the stdlib calls to `squin.h(q[0])`.

Also, there's probably a bug if you supply a custom list of `cirq`
qubits, since that's not passed into the `sub_frame` as far as I could
tell.

Instead of digging into it further for now, however, I decided to just
inline everything. Since the circuit we produce is flat anyway, the IR
shouldn't get too large. Of course, it's not ideal since we now inline
all the stdlib calls, but at least it works.