Add U3 gate statement to squin.gate (#534)
Closes #532 .

@weinbe58 @johnzl-777 Before merging, we need to clarify the convention
we are using for U3.

Both PyQrack and QASM2 use the definition

$$ U3(\theta, \phi, \lambda) = R_z(\phi)R_y(\theta)R_z(\lambda) =
R_y(\phi + \frac{\pi}{2})R_x(\theta)R_z(\lambda - \frac{\pi}{2}) $$

whereas qiskit and the stdlib we had before differ by a phase factor

$$ e^{(\phi + \lambda) / 2} R_z(\phi)R_y(\theta)R_z(\lambda) $$


Which one should we use? For now, I stuck with the first one, mainly
because I can just forward to the PyQrack `sim_reg.u` method without
modifications. We either need to change that, or the standard library
definitions. Since it's a statement, this should also be clearly
communicated, so I'd update the docstrings too, once we made a decision.