Fix SquinU3ToClifford Tests with Kirin 0.17.33 (#591)
Between Kirin 0.17.32 and .33 some changes introduced change the
behavior of CSE that caused some underlying assumptions about the
program structure to be changed and invalidated the old tests that
relied on statements at certain positions in the program. To be
specific, I believe the `none` statement that used to sit at the bottom
of the kernel now exists in the middle of the program (@weinbe58 was a
big help in figuring this out, I didn't even notice the unusual behavior
until looking again).

To be clear, the program is still correct and Phillip has pointed out
CSE probably latches onto the first instance of a statement as the one
to collapse all others to instead of the _last_ which is how it used to
be.

I've refactored the tests to be completely agnostic to the idea of
"statement at a certain index/position", focusing only on correct
gates/gate order (most conversions have more than one) This wasn't a
particularly good way to write tests and was a habit I acquired looking
at older Kirin source code 😅

---------

Co-authored-by: David Plankensteiner <david-pl@users.noreply.github.com>