Fix `CallGraphPass` issues in Kirin 0.20 (#586)
In this PR I port the `CallGraphPass` originally in Kirin 0.17 to the
passes here in bloqade-circuit. I have also updated them to use the new
Kirin APIs so they are not compatible with Kirin 0.20.

Unfortunately the tests are currently breaking because of some import
issues so the CI will not be able to pass for the tests until those are
fixed.

Co-authored-by: David Plankensteiner <david-pl@users.noreply.github.com>