New Address Analysis.  (#563)
~In this PR I implement a Joint analysis with constant prop and Address
analysis to try to get the address analysis to work properly going into
the call stack.~

I do not use a Cartesian product for the lattic because I need to
implement partial lambda and partial IList and tuple to get the analysis
to work. In order to support constant folding, however, I opted into
adding an extra lattice element to wrap the constant prop results.

---------

Co-authored-by: David Plankensteiner <david-pl@users.noreply.github.com>