Implement gemini.logical and draft kernel validation (#558)
This implements the simple dialect group as discussed in #531.

However, it also contains a draft (RFC) for validating kernels, using
the above as an example.

I thought a bit about it and in the end decided to split the validation
into two parts:

1. A `ValidationAnalysis` pass with an appropriate lattice (could be
extended to warnings, etc.) that checks for errors in the IR.
2. A simple `KernelValidation` class, that gathers up the errors and
throws them as `kirin.ir.ValidationError` with the appropriate source
info pointing to the line of the error. I went back and forth as to how
to implement this, since it is a bit similar to interpretation. However,
in the end I just decided to go with a very simple dataclass, since it
is sufficiently different from an interpreter or a rewrite.

One thing that's not great and I'm not sure how to improve: we probably
don't want the `ValidationAnalysis` to implement a lot of methods via
method tables. For example, the `gemini.logical` dialect doesn't support
`scf.IfElse`. However, if we want to re-use the `ValidationAnalysis`
elsewhere, then registering a `MethodTable` for `scf` with a method for
`scf.IfElse` might fail validation in kernels where this should be fine.
At the same time, we might want to share validation between different
kernels.

The way I "solved" this for now is to define a new analysis pass just
for the logical dialect that inherits from the generic
`ValidationAnalysis`. If we go this route, then the `ValidationAnalysis`
basically just defines the lattice and a fallback method.

There's also some TODOs left here, e.g. how to display multiple errors,
but I thought I'd get comments as soon as possible.

---------

Co-authored-by: Phillip Weinberg <pweinberg@quera.com>