Allow measurements and resets into Gemini noise model. (#583)
To simulate some operations, for example multiple rounds of syndrome
extraction, it is useful to be able to measure and reuse the measurement
qubits in the cirq simulation to save on RAM (even if on Gemini we will
actually use different measurement atoms for each round of syndrome
extraction).

Here is a minimum fix to allow for measurement and reset gates in the
noise models. The next step would be to add error in the measurement
that is proportional to the number of gate operations left in the
circuit, since in Gemini we will have to wait until the end of the
circuit to measure any atom.

@david-pl what do you think is the right course of action?

---------

Co-authored-by: David Plankensteiner <dplankensteiner@quera.com>
Co-authored-by: David Plankensteiner <david-pl@users.noreply.github.com>