Properly implement fidelity analysis with SCF (#640)
Closes #592 .

We are now mostly re-using the `AddressAnalysis` for walking the IR.
Except for `IfElse` statements, where we need to take care when merging
results from branches.
To do this, I had to make a small change to `AddressAnalysis`, but that
shouldn't affect anything else.

Probably needs more tests, but basics seem to work quite well.

Also, we might need some wrapper around running this analysis to have a
decent UX when trying to study the fidelity of a given program.

@weinbe58 I'm also not sure if I'm being to strict about expecting an
`AddressReg` for `qubtis: IList[Qubit]` arguments.

---------

Co-authored-by: Phillip Weinberg <pweinberg@quera.com>