TerminalLogicalMeasurement Validation (#641)
This is what I think should be sufficient for TerminalLogicalMeasurement
Validation.

I enforce the following conditions:
1. A program should only ever have ONE TerminalLogicalMeasurement
2. all spawned qubits should be consumed in that one terminal
measurement. There should be no dangling qubits.

There are some additions that come to mind - ranging from immediately
relevant to some things that are further out in terms of infrastructure:
- I had to make a separate validation pass as opposed to tacking things
onto the current logical validation. I think because of this "logical
validation" should be renamed to something a bit more precise.
- AddressAnalysis already counts the number of (unique) qubits spawned
but there's no way to access that information directly. It would be nice
if there was some way to attach that to the final frame returned at the
end of the analysis. Currently I tally up unique qubit ids and compare
that against the number of things fed into the terminal measurement

---------

Co-authored-by: Phillip Weinberg <pweinberg@quera.com>