Fix state initialization validation in gemini logical (#653)
Previously, we'd only allow any single non-Clifford gate to be applied
to a set of qubits. That means, you couldn't e.g. apply an `Rx` to the
first and an `Ry` to the second qubit, since the latter would have
already counted as the second gate and it's non-Clifford.

This fixes this by actually checking the first gates per qubit address.