QASM2 to squin (#644)
This is my implementation of #540 , there is one thing that I've held
off on after talking with @weinbe58 that's a bit tricky: handling
measurements. My understanding is we don't need it for now but in the
future it will be relevant.

While 95% of qasm2 maps very nicely to existing squin infrastructure
(with the exception of some of the less common 2 and 3 qubit gates I had
to skip - unless we want to do some implicit conversion for them) you
run into problems trying to map the qasm2 measurement semantics.

Take the following examples:

1. A user wants to measure out all the qubits to the entirety of the
declared register. Your qasm2 looks something like this
```
qreg q[5]
creg c[5]
measure q -> c
```
You could potentially just have creg be an empty ilist and then when the
measure happens produce the ilist of measurements - making sure that all
subsequent references point to the new ilist. This seems acceptable but
it gets tricky if a user wants to do something like

2. a user only wants to measure a single qubit and save that to a _slot_
in the classical register. This looks like

```
qreg q[5]
creg c[5]
measure q[0] -> c[0]
```
Now the classical register somehow has to already be `get_item`-able
which is tricky considering there's no notion of a fixed-length list.
Things get even trickier if a user wants to save at a different index:

```
measure q[1] -> c[3]
```

P.S. The sub kernel rewrite seems "off" to me, I'm sure there must be a
better way to do it but I'm not aware of what the proper API for it is 😅

---------

Co-authored-by: Phillip Weinberg <pweinberg@quera.com>